refactor(storage): fold sqlite and postgres onto a shared SQL core
sqlite and postgres now embed one shared sqlStore implementation and
differ only in placeholder style (rebind ?->$N). postgres adopts the
same single-table, embedded-messages schema that sqlite and d1 already
share, so a single SQL core drives every SQL backend and per-backend
files reduce to connection/dialect glue.

This also repairs the postgres list queries, which selected a
non-existent `summary` column and mismatched the row scan and had no
happy-path coverage. A shared conformance suite now runs against sqlite,
d1, and (gated on INFER_TEST_POSTGRES_DSN) a real postgres, wired into CI
via a postgres service container.

Closes #839
