+++ internal/infra/storage/d1_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"database/sql"
 	"encoding/json"
-	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -13,8 +12,6 @@ import (
 	"testing"
 	"time"
 
-	domain "github.com/inference-gateway/cli/internal/domain"
-	sdk "github.com/inference-gateway/sdk"
 	assert "github.com/stretchr/testify/assert"
 	require "github.com/stretchr/testify/require"
 	_ "modernc.org/sqlite"
@@ -161,230 +158,13 @@ func setupTestD1Storage(t *testing.T) *D1Storage {
 	return storage
 }
 
-func TestD1Storage_BasicOperations(t *testing.T) {
-	storage := setupTestD1Storage(t)
-	ctx := context.Background()
-
-	t.Run("Health Check", func(t *testing.T) {
-		assert.NoError(t, storage.Health(ctx))
+// TestD1Storage_Conformance runs the shared storage suite against the D1 driver,
+// backed by the httptest mock over real SQLite (unconditional in CI). D1 uses
+// the same single-table schema as SQLite/Postgres (see #839).
+func TestD1Storage_Conformance(t *testing.T) {
+	runConversationStorageConformance(t, func(t *testing.T) ConversationStorage {
+		return setupTestD1Storage(t)
 	})
-
-	t.Run("Save and Load Conversation", func(t *testing.T) {
-		conversationID := "test-conversation-1"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		require.NoError(t, storage.SaveConversation(ctx, conversationID, entries, metadata))
-
-		loadedEntries, loadedMetadata, err := storage.LoadConversation(ctx, conversationID)
-		require.NoError(t, err)
-
-		assert.Equal(t, metadata.ID, loadedMetadata.ID)
-		assert.Equal(t, metadata.Title, loadedMetadata.Title)
-		assert.Equal(t, len(entries), loadedMetadata.MessageCount)
-		assert.Equal(t, metadata.TokenStats, loadedMetadata.TokenStats)
-		assert.Equal(t, metadata.Tags, loadedMetadata.Tags)
-
-		assert.Len(t, loadedEntries, len(entries))
-		for i, entry := range entries {
-			assert.Equal(t, entry.Message.Content, loadedEntries[i].Message.Content)
-			assert.Equal(t, entry.Message.Role, loadedEntries[i].Message.Role)
-			assert.Equal(t, entry.Model, loadedEntries[i].Model)
-			assert.Equal(t, entry.Hidden, loadedEntries[i].Hidden)
-		}
-	})
-
-	t.Run("Update Conversation", func(t *testing.T) {
-		conversationID := "test-conversation-update"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		require.NoError(t, storage.SaveConversation(ctx, conversationID, entries, metadata))
-
-		newEntry := domain.ConversationEntry{
-			Message: sdk.Message{
-				Role:    sdk.Assistant,
-				Content: sdk.NewMessageContent("Updated response"),
-			},
-			Time:  time.Now(),
-			Model: "claude-4",
-		}
-		entries = append(entries, newEntry)
-
-		metadata.Title = "Updated Title"
-		metadata.UpdatedAt = time.Now()
-		metadata.MessageCount = len(entries)
-
-		require.NoError(t, storage.SaveConversation(ctx, conversationID, entries, metadata))
-
-		loadedEntries, loadedMetadata, err := storage.LoadConversation(ctx, conversationID)
-		require.NoError(t, err)
-
-		assert.Equal(t, "Updated Title", loadedMetadata.Title)
-		assert.Len(t, loadedEntries, len(entries))
-		lastContent, _ := loadedEntries[len(loadedEntries)-1].Message.Content.AsMessageContent0()
-		assert.Equal(t, "Updated response", lastContent)
-	})
-}
-
-func TestD1Storage_ConversationManagement(t *testing.T) {
-	storage := setupTestD1Storage(t)
-	ctx := context.Background()
-
-	t.Run("List Conversations", func(t *testing.T) {
-		conversations := []string{"conv1", "conv2", "conv3"}
-
-		for i, id := range conversations {
-			entries := createTestEntries()
-			metadata := createTestMetadata(id)
-			metadata.Title = "Conversation " + string(rune('A'+i))
-			metadata.CreatedAt = time.Now().Add(time.Duration(i) * time.Hour)
-			metadata.UpdatedAt = metadata.CreatedAt
-
-			require.NoError(t, storage.SaveConversation(ctx, id, entries, metadata))
-		}
-
-		summaries, err := storage.ListConversations(ctx, 10, 0)
-		require.NoError(t, err)
-		assert.GreaterOrEqual(t, len(summaries), 3)
-
-		for i := 1; i < len(summaries); i++ {
-			assert.True(t, summaries[i-1].UpdatedAt.After(summaries[i].UpdatedAt) ||
-				summaries[i-1].UpdatedAt.Equal(summaries[i].UpdatedAt))
-		}
-	})
-
-	t.Run("Delete Conversation", func(t *testing.T) {
-		conversationID := "test-conversation-delete"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		require.NoError(t, storage.SaveConversation(ctx, conversationID, entries, metadata))
-
-		_, _, err := storage.LoadConversation(ctx, conversationID)
-		require.NoError(t, err)
-
-		require.NoError(t, storage.DeleteConversation(ctx, conversationID))
-
-		_, _, err = storage.LoadConversation(ctx, conversationID)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "conversation not found")
-	})
-
-	t.Run("Update Metadata", func(t *testing.T) {
-		conversationID := "test-conversation-metadata"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		require.NoError(t, storage.SaveConversation(ctx, conversationID, entries, metadata))
-
-		metadata.Title = "New Title"
-		metadata.Tags = []string{"updated", "test"}
-		metadata.UpdatedAt = time.Now()
-
-		require.NoError(t, storage.UpdateConversationMetadata(ctx, conversationID, metadata))
-
-		_, loadedMetadata, err := storage.LoadConversation(ctx, conversationID)
-		require.NoError(t, err)
-
-		assert.Equal(t, "New Title", loadedMetadata.Title)
-		assert.Equal(t, []string{"updated", "test"}, loadedMetadata.Tags)
-	})
-}
-
-func TestD1Storage_ErrorCases(t *testing.T) {
-	storage := setupTestD1Storage(t)
-	ctx := context.Background()
-
-	t.Run("Load Non-existent Conversation", func(t *testing.T) {
-		_, _, err := storage.LoadConversation(ctx, "non-existent")
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "conversation not found")
-	})
-
-	t.Run("Delete Non-existent Conversation", func(t *testing.T) {
-		err := storage.DeleteConversation(ctx, "non-existent")
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "conversation not found")
-	})
-}
-
-// TestD1Storage_ListConversationsNeedingTitles guards the issue's explicit
-// warning: the title-generation batch path must receive Model/Tags/RequestCount
-// (i.e. the full mapper, not the lean ListConversations one).
-func TestD1Storage_ListConversationsNeedingTitles(t *testing.T) {
-	storage := setupTestD1Storage(t)
-	ctx := context.Background()
-
-	for i := range 3 {
-		id := fmt.Sprintf("needs-title-%d", i)
-		entries := createTestEntries() // 4 entries → count >= 2
-		metadata := createTestMetadata(id)
-		metadata.TitleGenerated = false
-		metadata.UpdatedAt = time.Now().Add(time.Duration(i) * time.Hour)
-		require.NoError(t, storage.SaveConversation(ctx, id, entries, metadata))
-	}
-
-	summaries, err := storage.ListConversationsNeedingTitles(ctx, 10)
-	require.NoError(t, err)
-	require.GreaterOrEqual(t, len(summaries), 3)
-
-	for _, s := range summaries {
-		assert.Equal(t, "claude-4", s.Model, "Model must be carried for the title batch path")
-		assert.Equal(t, []string{"test", "demo"}, s.Tags, "Tags must be carried for the title batch path")
-		assert.Equal(t, s.MessageCount/2, s.TokenStats.RequestCount)
-		assert.False(t, s.TitleGenerated)
-	}
-}
-
-func TestD1Storage_SessionGroups(t *testing.T) {
-	s := setupTestD1Storage(t)
-	ctx := context.Background()
-
-	_, ok, err := s.GetSessionGroup(ctx, "missing")
-	require.NoError(t, err)
-	assert.False(t, ok, "missing key must report not-found")
-
-	now := time.Now().UTC().Truncate(time.Second)
-	entry := SessionGroupEntry{
-		CurrentSessionID: "uuid-1",
-		History:          []string{"prev-a", "prev-b"},
-		LastRollover:     now,
-		UpdatedAt:        now,
-	}
-	require.NoError(t, s.PutSessionGroup(ctx, "channel-telegram-42", entry))
-
-	got, ok, err := s.GetSessionGroup(ctx, "channel-telegram-42")
-	require.NoError(t, err)
-	require.True(t, ok)
-	assert.Equal(t, "uuid-1", got.CurrentSessionID)
-	assert.Equal(t, []string{"prev-a", "prev-b"}, got.History)
-	assert.WithinDuration(t, now, got.UpdatedAt, time.Second)
-	assert.WithinDuration(t, now, got.LastRollover, time.Second)
-
-	require.NoError(t, s.PutSessionGroup(ctx, "channel-telegram-42", SessionGroupEntry{
-		CurrentSessionID: "uuid-2",
-		History:          []string{"prev-a", "prev-b", "uuid-1"},
-		UpdatedAt:        now,
-	}))
-
-	got, ok, err = s.GetSessionGroup(ctx, "channel-telegram-42")
-	require.NoError(t, err)
-	require.True(t, ok)
-	assert.Equal(t, "uuid-2", got.CurrentSessionID)
-	assert.Equal(t, []string{"prev-a", "prev-b", "uuid-1"}, got.History)
-	assert.True(t, got.LastRollover.IsZero(), "LastRollover must be cleared when UPSERT supplies zero value")
-
-	require.NoError(t, s.PutSessionGroup(ctx, "second", SessionGroupEntry{
-		CurrentSessionID: "uuid-3",
-		UpdatedAt:        now,
-	}))
-	all, err := s.ListSessionGroups(ctx)
-	require.NoError(t, err)
-	assert.Len(t, all, 2)
-	assert.Equal(t, "uuid-2", all["channel-telegram-42"].CurrentSessionID)
-	assert.Equal(t, "uuid-3", all["second"].CurrentSessionID)
-	assert.True(t, all["second"].LastRollover.IsZero())
 }
 
 // TestD1Storage_RequestShape asserts the driver hits the documented D1 endpoint
