+++ internal/infra/storage/sqlite_test.go
@@ -1,15 +1,10 @@
 package storage
 
 import (
-	"context"
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
-	domain "github.com/inference-gateway/cli/internal/domain"
-	sdk "github.com/inference-gateway/sdk"
-	assert "github.com/stretchr/testify/assert"
 	require "github.com/stretchr/testify/require"
 )
 
@@ -19,11 +14,7 @@ func setupTestStorage(t *testing.T) (*SQLiteStorage, func()) {
 
 	dbPath := filepath.Join(tempDir, "test.db")
 
-	config := SQLiteConfig{
-		Path: dbPath,
-	}
-
-	storage, err := NewSQLiteStorage(config)
+	storage, err := NewSQLiteStorage(SQLiteConfig{Path: dbPath})
 	require.NoError(t, err)
 
 	cleanup := func() {
@@ -34,273 +25,12 @@ func setupTestStorage(t *testing.T) (*SQLiteStorage, func()) {
 	return storage, cleanup
 }
 
-func TestSQLiteStorage_BasicOperations(t *testing.T) {
-	storage, cleanup := setupTestStorage(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("Health Check", func(t *testing.T) {
-		err := storage.Health(ctx)
-		assert.NoError(t, err)
+// TestSQLiteStorage_Conformance runs the shared storage suite against a real
+// on-disk SQLite database (pure-Go driver, unconditional in CI).
+func TestSQLiteStorage_Conformance(t *testing.T) {
+	runConversationStorageConformance(t, func(t *testing.T) ConversationStorage {
+		storage, cleanup := setupTestStorage(t)
+		t.Cleanup(cleanup)
+		return storage
 	})
-
-	t.Run("Save and Load Conversation", func(t *testing.T) {
-		conversationID := "test-conversation-1"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		err := storage.SaveConversation(ctx, conversationID, entries, metadata)
-		assert.NoError(t, err)
-
-		loadedEntries, loadedMetadata, err := storage.LoadConversation(ctx, conversationID)
-		assert.NoError(t, err)
-
-		assert.Equal(t, metadata.ID, loadedMetadata.ID)
-		assert.Equal(t, metadata.Title, loadedMetadata.Title)
-		assert.Equal(t, len(entries), loadedMetadata.MessageCount)
-		assert.Equal(t, metadata.TokenStats, loadedMetadata.TokenStats)
-		assert.Equal(t, metadata.Tags, loadedMetadata.Tags)
-
-		assert.Len(t, loadedEntries, len(entries))
-		for i, entry := range entries {
-			assert.Equal(t, entry.Message.Content, loadedEntries[i].Message.Content)
-			assert.Equal(t, entry.Message.Role, loadedEntries[i].Message.Role)
-			assert.Equal(t, entry.Model, loadedEntries[i].Model)
-			assert.Equal(t, entry.Hidden, loadedEntries[i].Hidden)
-		}
-	})
-
-	t.Run("Update Conversation", func(t *testing.T) {
-		conversationID := "test-conversation-update"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		err := storage.SaveConversation(ctx, conversationID, entries, metadata)
-		assert.NoError(t, err)
-
-		newEntry := domain.ConversationEntry{
-			Message: sdk.Message{
-				Role:    sdk.Assistant,
-				Content: sdk.NewMessageContent("Updated response"),
-			},
-			Time:  time.Now(),
-			Model: "claude-4",
-		}
-		entries = append(entries, newEntry)
-
-		metadata.Title = "Updated Title"
-		metadata.UpdatedAt = time.Now()
-		metadata.MessageCount = len(entries)
-
-		err = storage.SaveConversation(ctx, conversationID, entries, metadata)
-		assert.NoError(t, err)
-
-		loadedEntries, loadedMetadata, err := storage.LoadConversation(ctx, conversationID)
-		assert.NoError(t, err)
-
-		assert.Equal(t, "Updated Title", loadedMetadata.Title)
-		assert.Len(t, loadedEntries, len(entries))
-		lastContent, _ := loadedEntries[len(loadedEntries)-1].Message.Content.AsMessageContent0()
-		assert.Equal(t, "Updated response", lastContent)
-	})
-}
-
-func TestSQLiteStorage_ConversationManagement(t *testing.T) {
-	storage, cleanup := setupTestStorage(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("List Conversations", func(t *testing.T) {
-		conversations := []string{"conv1", "conv2", "conv3"}
-
-		for i, id := range conversations {
-			entries := createTestEntries()
-			metadata := createTestMetadata(id)
-			metadata.Title = "Conversation " + string(rune('A'+i))
-			metadata.CreatedAt = time.Now().Add(time.Duration(i) * time.Hour) // Different times
-			metadata.UpdatedAt = metadata.CreatedAt
-
-			err := storage.SaveConversation(ctx, id, entries, metadata)
-			require.NoError(t, err)
-		}
-
-		summaries, err := storage.ListConversations(ctx, 10, 0)
-		assert.NoError(t, err)
-		assert.GreaterOrEqual(t, len(summaries), 3)
-
-		for i := 1; i < len(summaries); i++ {
-			assert.True(t, summaries[i-1].UpdatedAt.After(summaries[i].UpdatedAt) ||
-				summaries[i-1].UpdatedAt.Equal(summaries[i].UpdatedAt))
-		}
-	})
-
-	t.Run("Delete Conversation", func(t *testing.T) {
-		conversationID := "test-conversation-delete"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		err := storage.SaveConversation(ctx, conversationID, entries, metadata)
-		assert.NoError(t, err)
-
-		_, _, err = storage.LoadConversation(ctx, conversationID)
-		assert.NoError(t, err)
-
-		err = storage.DeleteConversation(ctx, conversationID)
-		assert.NoError(t, err)
-
-		_, _, err = storage.LoadConversation(ctx, conversationID)
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "conversation not found")
-	})
-
-	t.Run("Update Metadata", func(t *testing.T) {
-		conversationID := "test-conversation-metadata"
-		entries := createTestEntries()
-		metadata := createTestMetadata(conversationID)
-
-		err := storage.SaveConversation(ctx, conversationID, entries, metadata)
-		assert.NoError(t, err)
-
-		metadata.Title = "New Title"
-		metadata.Tags = []string{"updated", "test"}
-		metadata.UpdatedAt = time.Now()
-
-		err = storage.UpdateConversationMetadata(ctx, conversationID, metadata)
-		assert.NoError(t, err)
-
-		_, loadedMetadata, err := storage.LoadConversation(ctx, conversationID)
-		assert.NoError(t, err)
-
-		assert.Equal(t, "New Title", loadedMetadata.Title)
-		assert.Equal(t, []string{"updated", "test"}, loadedMetadata.Tags)
-	})
-}
-
-func TestSQLiteStorage_ErrorCases(t *testing.T) {
-	storage, cleanup := setupTestStorage(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	t.Run("Load Non-existent Conversation", func(t *testing.T) {
-		_, _, err := storage.LoadConversation(ctx, "non-existent")
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "conversation not found")
-	})
-
-	t.Run("Delete Non-existent Conversation", func(t *testing.T) {
-		err := storage.DeleteConversation(ctx, "non-existent")
-		assert.Error(t, err)
-		assert.Contains(t, err.Error(), "conversation not found")
-	})
-}
-
-func createTestEntries() []domain.ConversationEntry {
-	now := time.Now()
-	return []domain.ConversationEntry{
-		{
-			Message: sdk.Message{
-				Role:    sdk.User,
-				Content: sdk.NewMessageContent("Hello, world!"),
-			},
-			Time:  now,
-			Model: "claude-4",
-		},
-		{
-			Message: sdk.Message{
-				Role:    sdk.Assistant,
-				Content: sdk.NewMessageContent("Hello! How can I help you today?"),
-			},
-			Time:  now.Add(time.Second),
-			Model: "claude-4",
-		},
-		{
-			Message: sdk.Message{
-				Role:    sdk.User,
-				Content: sdk.NewMessageContent("What is the capital of France?"),
-			},
-			Time:  now.Add(2 * time.Second),
-			Model: "claude-4",
-		},
-		{
-			Message: sdk.Message{
-				Role:    sdk.Assistant,
-				Content: sdk.NewMessageContent("The capital of France is Paris."),
-			},
-			Time:  now.Add(3 * time.Second),
-			Model: "claude-4",
-		},
-	}
-}
-
-func createTestMetadata(id string) ConversationMetadata {
-	now := time.Now()
-	return ConversationMetadata{
-		ID:        id,
-		Title:     "Test Conversation",
-		CreatedAt: now,
-		UpdatedAt: now,
-		TokenStats: domain.SessionTokenStats{
-			TotalInputTokens:  100,
-			TotalOutputTokens: 150,
-			TotalTokens:       250,
-			RequestCount:      2,
-		},
-		Model: "claude-4",
-		Tags:  []string{"test", "demo"},
-	}
-}
-
-func TestSQLiteStorage_SessionGroups(t *testing.T) {
-	s, cleanup := setupTestStorage(t)
-	defer cleanup()
-
-	ctx := context.Background()
-
-	_, ok, err := s.GetSessionGroup(ctx, "missing")
-	require.NoError(t, err)
-	assert.False(t, ok, "missing key must report not-found")
-
-	now := time.Now().UTC().Truncate(time.Second)
-	entry := SessionGroupEntry{
-		CurrentSessionID: "uuid-1",
-		History:          []string{"prev-a", "prev-b"},
-		LastRollover:     now,
-		UpdatedAt:        now,
-	}
-	require.NoError(t, s.PutSessionGroup(ctx, "channel-telegram-42", entry))
-
-	got, ok, err := s.GetSessionGroup(ctx, "channel-telegram-42")
-	require.NoError(t, err)
-	require.True(t, ok)
-	assert.Equal(t, "uuid-1", got.CurrentSessionID)
-	assert.Equal(t, []string{"prev-a", "prev-b"}, got.History)
-	assert.WithinDuration(t, now, got.UpdatedAt, time.Second)
-	assert.WithinDuration(t, now, got.LastRollover, time.Second)
-
-	require.NoError(t, s.PutSessionGroup(ctx, "channel-telegram-42", SessionGroupEntry{
-		CurrentSessionID: "uuid-2",
-		History:          []string{"prev-a", "prev-b", "uuid-1"},
-		UpdatedAt:        now,
-	}))
-
-	got, ok, err = s.GetSessionGroup(ctx, "channel-telegram-42")
-	require.NoError(t, err)
-	require.True(t, ok)
-	assert.Equal(t, "uuid-2", got.CurrentSessionID)
-	assert.Equal(t, []string{"prev-a", "prev-b", "uuid-1"}, got.History)
-	assert.True(t, got.LastRollover.IsZero(), "LastRollover must be cleared when UPSERT supplies zero value")
-
-	require.NoError(t, s.PutSessionGroup(ctx, "second", SessionGroupEntry{
-		CurrentSessionID: "uuid-3",
-		UpdatedAt:        now,
-	}))
-	all, err := s.ListSessionGroups(ctx)
-	require.NoError(t, err)
-	assert.Len(t, all, 2)
-	assert.Equal(t, "uuid-2", all["channel-telegram-42"].CurrentSessionID)
-	assert.Equal(t, "uuid-3", all["second"].CurrentSessionID)
-	assert.True(t, all["second"].LastRollover.IsZero())
 }
